update code call api current day and focast 5days turn 2

+++ app/src/main/java/com/example/weatheronline/model/weatherresult/Maximum.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 class Maximum(
     @SerializedName("Value")
     @Expose
-    val value: Int? = null,
+    val value: Float? = null,
 
     @SerializedName("Unit")
     @Expose
